fix(presets): refine ai tracking properties (#6966)
- refine ai tracking properties
- fix signal handling
