+++ packages/presets/src/ai/index.ts
@@ -4,4 +4,3 @@ export * from './entries/edgeless/actions-config.js';
 export * from './entries/index.js';
 export * from './messages/index.js';
 export * from './provider.js';
-export * from './utils/event-source.js';
